pyproject.toml
[tool.poetry]
name = "river-api"
version = "0.1.0"
description = "API to manage data for RIVER lab"
authors = ["Evan Thomas <evan.thomas@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
aiosqlite = "^0.19.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.21.1"
asyncpg = "^0.29.0"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.12.1"
setuptools = "^68.2.2"
wheel = "^0.41.3"
numpy = "^1.26.1"
requests = "^2.31.0"
aiohttp = "^3.9.1"
tenacity = "^8.2.3"
pydantic-settings = "^2.1.0"
sqlmodel-react-admin = "^0.0.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-watcher = "^0.3.4"
pytest-asyncio = "^0.21.1"
bump-pydantic = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
